Refactoring: split kalman_filter.py into a several files in a folder
